Fix/codeql main alerts

+++ examples/radump/print-syslog.c
@@ -89,7 +89,7 @@ static const struct tok syslog_facility_values[] = {
 char *
 syslog_print(register const u_char *pptr, register u_int len)
 {
-    u_int16_t msg_off = 0;
+    u_int msg_off = 0;
     u_int16_t pri = 0;
     u_int16_t facility,severity;
